Runtime Catch & Manual Creature List Addition Support

+++ code/modules/client/preferences.dm
@@ -564,10 +564,15 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 			else
 				dat += "[TextPreview(creature_ooc)]...<br>"
 			if(!creature_image && creature_species)
+				if(!LAZYLEN(GLOB.creature_selectable))//Pokemon selection list is empty, so generate it.
+					generate_selectable_creatures()
+				if(!(creature_species in GLOB.creature_selectable))//Previously selected species which isn't supported anymore.
+					creature_species = initial(creature_species)
 				var/creature_type = GLOB.creature_selectable["[creature_species]"]
-				var/mob/living/M = new creature_type(user)
-				creature_image = image(icon=M.icon,icon_state=M.icon_state,dir=2)
-				qdel(M)
+				if(!isnull(creature_type) && isliving(creature_type))//If we couldn't find a type to spawn, avoid a runtime and don't try to make a null
+					var/mob/living/M = new creature_type(user)
+					creature_image = image(icon=M.icon,icon_state=M.icon_state,dir=2)
+					qdel(M)
 			if(creature_image)
 				dat += "[icon2html(creature_image, user)]<br>"
 
@@ -2415,6 +2420,10 @@ GLOBAL_LIST_EMPTY(preferences_datums)
 						else
 							to_chat(user, "<font color='red'>Invalid name. Your name should be at least 2 and at most [MAX_NAME_LEN] characters long. It may only contain the characters A-Z, a-z, -, ' and .</font>")
 				if("creature_species")
+					if(!LAZYLEN(GLOB.creature_selectable))//Pokemon selection list is empty, so generate it.
+						generate_selectable_creatures()
+					if(!(creature_species in GLOB.creature_selectable))//Previously selected species which isn't supported anymore.
+						creature_species = initial(creature_species)
 					var/result = input(user, "Select a creature species", "Species Selection") as null|anything in GLOB.creature_selectable
 					if(result)
 						creature_species = result
